Factor out Interop APIs from netstandard ref
1. Remove TypeLib APIs as they aren't needed or compatible
(see https://github.com/dotnet/standard/issues/2).
2. Remove System.Runtime.Interop.ComTypes namespace as it is only
COM interfaces and structs which aren't applicable to all OS's and
we can easily make that a compat library as needed.
3. Removed all obsolete structs that were also in ComTypes.
4. Marshal type was sub-setted to match what Xamarin exposeds
(see https://github.com/dotnet/standard/issues/18).
