+++ netstandard/ref/System.cs
@@ -7978,107 +7978,6 @@ namespace System.Runtime.InteropServices
         public void Add() { }
         public void Remove() { }
     }
-    public partial class StandardOleMarshalObject : System.MarshalByRefObject
-    {
-        protected StandardOleMarshalObject() { }
-    }
-}
-namespace System.Runtime.InteropServices.ComTypes
-{
-    [System.FlagsAttribute]
-    public enum ADVF
-    {
-        ADVF_DATAONSTOP = 64,
-        ADVF_NODATA = 1,
-        ADVF_ONLYONCE = 4,
-        ADVF_PRIMEFIRST = 2,
-        ADVFCACHE_FORCEBUILTIN = 16,
-        ADVFCACHE_NOHANDLER = 8,
-        ADVFCACHE_ONSAVE = 32,
-    }
-    public enum DATADIR
-    {
-        DATADIR_GET = 1,
-        DATADIR_SET = 2,
-    }
-    [System.FlagsAttribute]
-    public enum DVASPECT
-    {
-        DVASPECT_CONTENT = 1,
-        DVASPECT_DOCPRINT = 8,
-        DVASPECT_ICON = 4,
-        DVASPECT_THUMBNAIL = 2,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct FORMATETC
-    {
-        public short cfFormat;
-        public System.Runtime.InteropServices.ComTypes.DVASPECT dwAspect;
-        public int lindex;
-        public System.IntPtr ptd;
-        public System.Runtime.InteropServices.ComTypes.TYMED tymed;
-    }
-    public partial interface IAdviseSink
-    {
-        void OnClose();
-        void OnDataChange(ref System.Runtime.InteropServices.ComTypes.FORMATETC format, ref System.Runtime.InteropServices.ComTypes.STGMEDIUM stgmedium);
-        void OnRename(System.Runtime.InteropServices.ComTypes.IMoniker moniker);
-        void OnSave();
-        void OnViewChange(int aspect, int index);
-    }
-    public partial interface IDataObject
-    {
-        int DAdvise(ref System.Runtime.InteropServices.ComTypes.FORMATETC pFormatetc, System.Runtime.InteropServices.ComTypes.ADVF advf, System.Runtime.InteropServices.ComTypes.IAdviseSink adviseSink, out int connection);
-        void DUnadvise(int connection);
-        int EnumDAdvise(out System.Runtime.InteropServices.ComTypes.IEnumSTATDATA enumAdvise);
-        System.Runtime.InteropServices.ComTypes.IEnumFORMATETC EnumFormatEtc(System.Runtime.InteropServices.ComTypes.DATADIR direction);
-        int GetCanonicalFormatEtc(ref System.Runtime.InteropServices.ComTypes.FORMATETC formatIn, out System.Runtime.InteropServices.ComTypes.FORMATETC formatOut);
-        void GetData(ref System.Runtime.InteropServices.ComTypes.FORMATETC format, out System.Runtime.InteropServices.ComTypes.STGMEDIUM medium);
-        void GetDataHere(ref System.Runtime.InteropServices.ComTypes.FORMATETC format, ref System.Runtime.InteropServices.ComTypes.STGMEDIUM medium);
-        int QueryGetData(ref System.Runtime.InteropServices.ComTypes.FORMATETC format);
-        void SetData(ref System.Runtime.InteropServices.ComTypes.FORMATETC formatIn, ref System.Runtime.InteropServices.ComTypes.STGMEDIUM medium, bool release);
-    }
-    public partial interface IEnumFORMATETC
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumFORMATETC newEnum);
-        int Next(int celt, System.Runtime.InteropServices.ComTypes.FORMATETC[] rgelt, int[] pceltFetched);
-        int Reset();
-        int Skip(int celt);
-    }
-    public partial interface IEnumSTATDATA
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumSTATDATA newEnum);
-        int Next(int celt, System.Runtime.InteropServices.ComTypes.STATDATA[] rgelt, int[] pceltFetched);
-        int Reset();
-        int Skip(int celt);
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct STATDATA
-    {
-        public System.Runtime.InteropServices.ComTypes.ADVF advf;
-        public System.Runtime.InteropServices.ComTypes.IAdviseSink advSink;
-        public int connection;
-        public System.Runtime.InteropServices.ComTypes.FORMATETC formatetc;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct STGMEDIUM
-    {
-        public object pUnkForRelease;
-        public System.Runtime.InteropServices.ComTypes.TYMED tymed;
-        public System.IntPtr unionmember;
-    }
-    [System.FlagsAttribute]
-    public enum TYMED
-    {
-        TYMED_ENHMF = 64,
-        TYMED_FILE = 2,
-        TYMED_GDI = 16,
-        TYMED_HGLOBAL = 1,
-        TYMED_ISTORAGE = 8,
-        TYMED_ISTREAM = 4,
-        TYMED_MFPICT = 32,
-        TYMED_NULL = 0,
-    }
 }
 namespace System.Runtime.Versioning
 {
