+++ netstandard/src/ApiCompatBaseline.net461.txt
@@ -1,5 +1,3 @@
 Compat issues with assembly netstandard:
 CannotAddAbstractMembers : Member 'System.IO.IsolatedStorage.IsolatedStorage.GetPermission(System.Security.PermissionSet)' is abstract in the implementation but is missing in the contract.
-InterfacesShouldHaveSameMembers : Implementation interface member 'System.Runtime.InteropServices.ITypeLibConverter.ConvertTypeLibToAssembly(System.Object, System.String, System.Int32, System.Runtime.InteropServices.ITypeLibImporterNotifySink, System.Byte[], System.Reflection.StrongNameKeyPair, System.Boolean)' is not in the contract.
-InterfacesShouldHaveSameMembers : Implementation interface member 'System.Runtime.InteropServices.ITypeLibConverter.ConvertTypeLibToAssembly(System.Object, System.String, System.Runtime.InteropServices.TypeLibImporterFlags, System.Runtime.InteropServices.ITypeLibImporterNotifySink, System.Byte[], System.Reflection.StrongNameKeyPair, System.String, System.Version)' is not in the contract.
-Total Issues: 3
+Total Issues: 1
